+++ apps/backgrounds/app/layout.tsx
@@ -5,8 +5,9 @@ import "./globals.css";
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "Backgrounds By Grida",
+  description:
+    "A collection of beautiful backgrounds ready to use with iframe embedding.",
 };
 
 export default function RootLayout({
