+++ apps/backgrounds/package.json
@@ -9,18 +9,28 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@react-three/drei": "^9.83.9",
+    "@react-three/fiber": "^8.16.7",
+    "clsx": "^2.1.1",
+    "framer-motion": "^11.3.8",
+    "mini-svg-data-uri": "^1.4.4",
+    "next": "14.2.5",
     "react": "^18",
     "react-dom": "^18",
-    "next": "14.2.5"
+    "shadergradient": "^1.2.14",
+    "tailwind-merge": "^2.2.2",
+    "three": "^0.150.1",
+    "three-globe": "^2.31.1"
   },
   "devDependencies": {
-    "typescript": "^5",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
+    "@types/three": "^0.149.0",
+    "eslint": "^8",
+    "eslint-config-next": "14.2.5",
     "postcss": "^8",
     "tailwindcss": "^3.4.1",
-    "eslint": "^8",
-    "eslint-config-next": "14.2.5"
+    "typescript": "^5"
   }
 }
