+++ apps/backgrounds/theme/backgrounds/aurora/index.tsx
