--- apps/forms/app/(index)/wwwembeddings/beams/page.tsx
@@ -1,5 +1,0 @@
-import { BackgroundBeams } from "@/theme/backgrounds/beams";
-
-export default function BeamsBgPage() {
-  return <BackgroundBeams />;
-}
