--- apps/forms/app/(theme)/theme/embed/backgrounds/aurora/page.tsx
@@ -1,5 +1,0 @@
-import { AuroraBackground } from "@/theme/backgrounds/aurora";
-
-export default function AuroraBgPage() {
-  return <AuroraBackground />;
-}
