+++ apps/forms/package.json
@@ -18,8 +18,6 @@
     "@dnd-kit/modifiers": "^7.0.0",
     "@dnd-kit/sortable": "^8.0.0",
     "@dnd-kit/utilities": "^3.2.2",
-    "@editor-ui/alert-dialog": "workspace:^",
-    "@editor-ui/dropdown-menu": "workspace:^",
     "@faker-js/faker": "^8.4.1",
     "@fingerprintjs/fingerprintjs": "^4.2.2",
     "@floating-ui/react": "^0.26.10",
@@ -53,8 +51,6 @@
     "@radix-ui/react-toolbar": "^1.0.3",
     "@radix-ui/react-tooltip": "^1.0.7",
     "@react-email/components": "^0.0.18",
-    "@react-three/drei": "^9.101.0",
-    "@react-three/fiber": "^8.16.6",
     "@supabase/auth-helpers-nextjs": "^0.9.0",
     "@supabase/postgrest-js": "^1.15.5",
     "@supabase/ssr": "^0.1.0",
@@ -65,7 +61,6 @@
     "@tiptap/starter-kit": "^2.4.0",
     "@tosspayments/payment-widget-sdk": "^0.10.2",
     "@types/mdx": "^2.0.11",
-    "@types/three": "^0.149.0",
     "@use-gesture/react": "^10.2.11",
     "@uidotdev/usehooks": "^2.4.1",
     "@visx/responsive": "^3.10.2",
@@ -97,11 +92,10 @@
     "lucide-react": "^0.367.0",
     "mapbox-gl": "^3.4.0",
     "mdn-data": "^2.8.0",
-    "mini-svg-data-uri": "^1.4.4",
     "monaco-editor": "^0.47.0",
     "monaco-themes": "^0.4.4",
     "nanoid": "^3.1.23",
-    "next": "14.2.4",
+    "next": "14.2.5",
     "next-themes": "^0.3.0",
     "openai": "^4.29.1",
     "prism-react-renderer": "^2.3.1",
@@ -121,13 +115,10 @@
     "react-textarea-autosize": "^8.5.3",
     "react-timer-hook": "^3.0.7",
     "resend": "^3.2.0",
-    "shadergradient": "^1.2.14",
     "signature_pad": "^4.2.0",
     "stylis": "^4.3.2",
     "swr": "^2.2.5",
     "tailwind-merge": "^2.2.2",
-    "three": "^0.150.1",
-    "three-globe": "^2.31.0",
     "timezones-list": "^3.0.3",
     "use-file-picker": "^2.1.1",
     "usehooks-ts": "^3.1.0",
