+++ apps/forms/tailwind.config.ts
@@ -1,13 +1,5 @@
 import type { Config } from "tailwindcss";
 
-const svgToDataUri = require("mini-svg-data-uri");
-
-const colors = require("tailwindcss/colors");
-
-const {
-  default: flattenColorPalette,
-} = require("tailwindcss/lib/util/flattenColorPalette");
-
 const config: Config = {
   presets: [require("./tailwind.config.ui.js")],
   content: [
@@ -25,16 +17,6 @@ const config: Config = {
         DEFAULT: `rgb(var(--border-color))`,
       }),
       keyframes: {
-        // https://ui.aceternity.com/components/aurora-background
-        aurora: {
-          from: {
-            backgroundPosition: "50% 50%, 50% 50%",
-          },
-          to: {
-            backgroundPosition: "350% 50%, 350% 50%",
-          },
-        },
-        //
         fadeIn: {
           "0%": { transform: "scale(0.95)", opacity: "0" },
           "100%": { transform: "scale(1)", opacity: "1" },
@@ -138,9 +120,6 @@ const config: Config = {
         },
       },
       animation: {
-        // https://ui.aceternity.com/components/aurora-background
-        aurora: "aurora 60s linear infinite",
-        //
         "fade-in": "fadeIn 300ms both",
         "fade-out": "fadeOut 300ms both",
 
@@ -184,49 +163,7 @@ const config: Config = {
     },
     plugins: [require("tailwindcss-animate")],
   },
-  plugins: [
-    require("@tailwindcss/typography"),
-    addVariablesForColors,
-    /**
-     * https://ui.aceternity.com/components/grid-and-dot-backgrounds
-     */
-    function ({ matchUtilities, theme }: any) {
-      matchUtilities(
-        {
-          "bg-grid": (value: any) => ({
-            backgroundImage: `url("${svgToDataUri(
-              `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 32 32" width="32" height="32" fill="none" stroke="${value}"><path d="M0 .5H31.5V32"/></svg>`
-            )}")`,
-          }),
-          "bg-grid-small": (value: any) => ({
-            backgroundImage: `url("${svgToDataUri(
-              `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 32 32" width="8" height="8" fill="none" stroke="${value}"><path d="M0 .5H31.5V32"/></svg>`
-            )}")`,
-          }),
-          "bg-dot": (value: any) => ({
-            backgroundImage: `url("${svgToDataUri(
-              `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 32 32" width="16" height="16" fill="none"><circle fill="${value}" id="pattern-circle" cx="10" cy="10" r="1.6257413380501518"></circle></svg>`
-            )}")`,
-          }),
-        },
-        { values: flattenColorPalette(theme("backgroundColor")), type: "color" }
-      );
-    },
-  ],
+  plugins: [require("@tailwindcss/typography")],
 };
 
 export default config;
-
-/**
- * @see https://ui.aceternity.com/components/aurora-background
- */
-function addVariablesForColors({ addBase, theme }: any) {
-  let allColors = flattenColorPalette(theme("colors"));
-  let newVars = Object.fromEntries(
-    Object.entries(allColors).map(([key, val]) => [`--${key}`, val])
-  );
-
-  addBase({
-    ":root": newVars,
-  });
-}
